# Conflicts:
#	EliteJournalReader.Tests/EliteJournalReader.Tests.csproj
#	EliteJournalReader/Events/NavRouteEvent.cs

+++ EliteJournalReader/Events/PayBountiesEvent.cs
@@ -9,11 +9,11 @@ namespace EliteJournalReader.Events
 {
     //When written: when paying off bounties
     //Parameters:
-    //•	Amount: (total amount paid, including any broker fee)
-    //•	BrokerPercentage(present if paid via a Broker)
-    //•	AllFines: bool
-    //•	Faction
-    //•	ShipID
+    //ï¿½	Amount: (total amount paid, including any broker fee)
+    //ï¿½	BrokerPercentage(present if paid via a Broker)
+    //ï¿½	AllFines: bool
+    //ï¿½	Faction
+    //ï¿½	ShipID
     public class PayBountiesEvent : JournalEvent<PayBountiesEvent.PayBountiesEventArgs>
     {
         public PayBountiesEvent() : base("PayBounties") { }
@@ -24,7 +24,7 @@ namespace EliteJournalReader.Events
             public double? BrokerPercentage { get; set; }
             public string Faction { get; set; }
             public string Faction_Localised { get; set; }
-            public int ShipID { get; set; }
+            public long ShipID { get; set; }
         }
     }
 }
